+++ packages/components/src/list-auto-grid/src/index.vue
@@ -1,19 +1,24 @@
 <template>
-  <div ref="rootRef" class="mixte-list-auto-grid" :style="rootStyle">
+  <div ref="rootRef" class="mixte-list-auto-grid" :style="unref(rootStyle)">
     <div v-for="(item, index) in list" :key="index" style="overflow: hidden">
       <slot :item="item" :index="index" />
     </div>
   </div>
 </template>
 
-<script lang="ts" setup>
+<script
+  setup
+  lang="ts"
+  generic="T extends object[] | number"
+>
+  import { isVue2, toRef, unref, useSlots } from 'vue-demi';
   import { useAutoGrid } from './composables/useAutoGrid';
 
   // @unocss-ignore
 
   interface ListAutoGridProps {
     /** 列表数据 */
-    list?: Record<PropertyKey, any>[] | number;
+    list?: T;
     /**
      * 组件宽度 (单位: px)
      *  - 正常情况无需使用, 会自动获取组件宽度
@@ -29,9 +34,26 @@
     gapY?: number | `${number}`;
   }
 
-  const props = withDefaults(defineProps<ListAutoGridProps>(), {
-    list: () => [],
-  });
+  const props = defineProps<ListAutoGridProps>();
+
+  // 为了在 Vue3 中拥有更好的类型推导, 但是 Vue2 并不支持这个 API, 先这样兼容一下
+  let oldDefineSlots: any;
+
+  if (isVue2) { // @ts-expect-error
+    oldDefineSlots = globalThis.defineSlots; // @ts-expect-error
+    globalThis.defineSlots = useSlots;
+  }
+
+  // eslint-disable-next-line vue/define-macros-order
+  defineSlots<{
+    default: (props: { item: T extends object[] ? T[number] : number; index: number }) => any;
+  }>();
+
+  if (isVue2) { // @ts-expect-error
+    globalThis.defineSlots = oldDefineSlots;
+  }
+
+  const list = toRef(props, 'list');
 
   const {
     rootRef,